requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-dotenv 